Formatted code and added buzzer on encoder turn

+++ main/menu.h
@@ -16,15 +16,15 @@
 
 #define DEBOUNCE_DELAY 150
 
-// Uncomment to use a rotary encoder switch
-//#define ROTARY_SWITCH
-
 // How long button has to be held to be long-pressed
 #define LONG_PRESS_DURATION (500 - DEBOUNCE_DELAY)
 
 // Keeps small area at top and bottom for text display on scan menu
 #define BAR_Y_MIN 14
 #define BAR_Y_MAX 57
+
+// Use rotary encoder instead of buttons for navigation
+// #define ROTARY_SWITCH
 
 // Enum for different menus
 // Order is important from initMenus()
@@ -47,9 +47,9 @@ class Menu {
 public:
   Menu(uint8_t p_p, uint8_t s_p, uint8_t n_p, Settings *s, Buzzer *b, RX5808 *r, Api *a);
   void begin();
-  static Menu *instance; // Static pointer to the current Menu instance
-  static void encoderWrapper(); // Static ISR wrapper function
-  void doEncoder(); // Non-static method for handling the encoder
+  static Menu *instance;         // Static pointer to current Menu instance
+  static void encoderWrapper();  // Static ISR wrapper function
+  void doEncoder();              // Non-static method for handling encoder
   void handleButtons();
   void clearBuffer();
   void sendBuffer();
@@ -93,9 +93,9 @@ private:
   uint8_t select_pin;
   uint8_t next_pin;
 
-  volatile uint16_t dial_pos;
-  uint16_t last_dial_pos;
-  volatile int8_t encoder_state;
+  volatile int dial_pos;
+  int last_dial_pos;
+  volatile int encoder_state;
 
   // Used to handle long-pressing SELECT to go back
   unsigned long selectButtonPressTime;
@@ -106,11 +106,10 @@ private:
   RX5808 *module;
   Api *api;
 
-  // If using an OLED with an SH1106 chip then leave this be
-  // If using an OLED with an SSD1306 chip then comment out the SH1106 line and uncomment the SSD1306 line
+  // Uncomment line for required display chip
   U8G2_SH1106_128X64_NONAME_F_HW_I2C u8g2;
-  //U8G2_SSD1306_128X64_NONAME_F_HW_I2C u8g2;
-  //U8G2_SSD1309_128X64_NONAME0_F_HW_I2C u8g2;
+  // U8G2_SSD1306_128X64_NONAME_F_HW_I2C u8g2;
+  // U8G2_SSD1309_128X64_NONAME0_F_HW_I2C u8g2;
 };
 
 #endif
